+++ src/t5/layer_norm.rs
@@ -32,7 +32,7 @@ impl T5LayerNorm {
 
 impl Module for T5LayerNorm {
     fn forward(&self, x: &Tensor) -> Tensor {
-        let variance = x.pow(2f64).mean_dim(&[-1], true, Kind::Float);
+        let variance = x.pow_tensor_scalar(2f64).mean_dim(&[-1], true, Kind::Float);
         let x = x * (variance + self.epsilon).rsqrt();
         &self.weight * x
     }